+++ examples/03_native.zig
@@ -46,6 +46,8 @@ fn onWindowEvent(ctx: ?*anyopaque, event: zgpui.platform.WindowEvent) void {
         .a11y_press => {},
         .a11y_adjust => {},
         .a11y_set_value => {},
+        .a11y_replace_selected_text => {},
+        .a11y_set_selected_range => {},
     }
 }